pyproject.toml
[project]
name = "sonic-o1"
version = "0.1.0"
description = "SONIC-O1: Social Non-scripted Interaction Corpus for Fair Multi-form Multimodal Video Understandin"
readme = "README.md"
authors = [{name = "Ahmed Radwan", email = "ahmed.radwan@vectorinstitute.ai"}]
license = "MIT"
requires-python = ">=3.10,<3.13"
dependencies = [
    # Core dependencies
    "numpy>=1.19,<2.0.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.0",
    "tqdm>=4.65.0",
    "omegaconf>=2.3.0",
    # AI/ML APIs
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "google-genai>=1.36.06",
    # HuggingFace ecosystem - FIXED for transformers 4.57.3
    "datasets[vision,audio]>=3.4.0,<4.0.0",
    "transformers>=4.57.3,<4.58.0",
    "huggingface-hub>=0.16.0,<1.0.0",
    "tokenizers>=0.15.0,<1.0.0",
    "accelerate>=0.20.0",
    "qwen-omni-utils[decord]>=0.0.8",
    # PyTorch 2.5.x (stable)
    "torch>=2.5.0,<2.6.0",
    "torchvision>=0.20.0,<0.21.0",
    "torchaudio>=2.5.0,<2.6.0",
    # PyTorch Video
    "pytorchvideo",
    # Video/Audio downloading
    "yt-dlp>=2023.1.0",
    # Data handling
    "pyarrow>=12.0.0,<16.0.0",
    "fsspec[http]>=2023.10.0,<2025.0.0",
    "multiprocess>=0.70.15,<1.0.0",
    "dill>=0.3.7,<1.0.0",
    "xxhash>=3.0.0,<4.0.0",
    "aiohttp>=3.9.0,<4.0.0",
    # Video processing
    "av==12.3.0",
    "imageio>=2.25.0,<2.35.0",
    "imageio-ffmpeg>=0.4.7,<0.5.0",
    "moviepy>=1.0.3,<1.1.0",
    "decord>=0.6.0,<0.7.0",
    # Image/Video processing
    "pillow>=9.0.0,<11.0.0",
    "opencv-python-headless>=4.6.0,<4.10.0",
    # Audio processing - REQUIRED for Qwen2.5-Omni
    "librosa>=0.9.2,<0.11.0",
    "soundfile>=0.11.0,<0.13.0",
    "scipy>=1.9.0,<1.14.0",
    # Video codec support
    "ffmpeg-python>=0.2.0",
    # ImageBind dependencies
    "timm",
    "ftfy",
    "regex",
    "einops",
    "iopath",
    "types-regex",
    # Additional ML utilities
    "scikit-learn>=1.1.0,<1.5.0",
    "matplotlib>=3.5.0,<3.9.0",
    "seaborn>=0.11.0,<0.14.0",
    # Video dataset handling
    "webdataset>=0.2.0,<0.3.0",
    # Additional dependencies
    "packaging>=21.0",
    "filelock>=3.12.0",
    "python-dotenv>=1.1.1",
    "google-api-python-client>=2.184.0",
    "youtube-transcript-api>=1.2.2",
    "rouge-score>=0.1.2",
    "pycocoevalcap>=1.2",
    "sentence-transformers>=5.2.0",
    "openpyxl>=3.1.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.9.2",
    "jupyter>=1.0.0",
    "ipython>=8.12.0",
    "ipykernel>=6.0.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv]
default-groups = ["dev"]
index-strategy = "first-index"

[tool.uv.sources]
pytorchvideo = { git = "https://github.com/facebookresearch/pytorchvideo.git" }

[tool.hatch.build.targets.wheel]
packages = ["sonic-o1"]
